+++ FiatChampAddon/FiatClient/AppConfig.cs
@@ -20,11 +20,12 @@ public record AppConfig
   public FcaBrand Brand { get; set; }
   public FcaRegion Region { get; set; } = FcaRegion.Europe;
   public string HomeAssistantUrl { get; set; } = "http://supervisor/core";
-  public int StartDelaySeconds { get; set; } = 1; 
+  public int StartDelaySeconds { get; set; } = 1;
   public bool AutoRefreshLocation { get; set; } = false;
   public bool AutoRefreshBattery { get; set; } = false;
   public bool EnableDangerousCommands { get; set; } = false;
   public bool ConvertKmToMiles { get; set; } = false;
+  public bool PrefixEntityNamesWithCarName { get; set; } = false;
   public bool DevMode { get; set; } = false;
   public bool UseFakeApi { get; set; } = false;
   public bool Debug { get; set; } = false;
@@ -44,7 +45,7 @@ public record AppConfig
 
     return JsonConvert.SerializeObject(tmp, Formatting.Indented, new StringEnumConverter());
   }
-  
+
   public bool IsPinSet()
   {
     return !string.IsNullOrWhiteSpace(this.FiatPin);